Finish propagating changes from commit af16f3
In such commit the `erase_filepath` parameter was introduced, but the change was only applied for the case when the full HDF5 file is read, and not for the case when the number of required waveforms by the user (given by the `wvfm_count` parameter) is reached midway through the reading process and it is interrupted.

+++ src/waffles/input_output/raw_hdf5_reader.py
@@ -491,7 +491,8 @@ def WaveformSet_from_hdf5_file(filepath : str,
                                     minimum_length
                                 )
 
-                        os.remove(filepath)
+                        if erase_filepath:
+                            os.remove(filepath)
                         return WaveformSet(*waveforms)
                     
     if truncate_wfs_to_minimum:
